WIP loplugin:droplong in tools/gen.hxx and vcl
large patch because the gen.hxx stuff just keeps propogating
everywhere.

+++ framework/source/layoutmanager/layoutmanager.cxx
@@ -2344,7 +2344,7 @@ bool LayoutManager::implts_doLayout( bool bForceRequestBorderSpace, bool bOuterR
             // Position the status bar
             if ( aStatusBarSize.Height() > 0 )
             {
-                implts_setStatusBarPosSize( ::Point( 0, std::max(( aContainerSize.Height() ), long( 0 ))),
+                implts_setStatusBarPosSize( ::Point( 0, std::max(aContainerSize.Height(), 0)),
                                             ::Size( aContainerSize.Width(),aStatusBarSize.Height() ));
             }
 
@@ -2467,7 +2467,7 @@ void LayoutManager::implts_setDockingAreaWindowSizes()
     // Position the status bar
     if ( aStatusBarSize.Height() > 0 )
     {
-        implts_setStatusBarPosSize( ::Point( 0, std::max(( aContainerClientSize.Height - aStatusBarSize.Height() ), long( 0 ))),
+        implts_setStatusBarPosSize( ::Point( 0, std::max(( aContainerClientSize.Height - aStatusBarSize.Height() ), 0)),
                                     ::Size( aContainerClientSize.Width, aStatusBarSize.Height() ));
     }
 }
